Add servise for Every Route
