+++ services/doctorService.js
@@ -1,0 +1,19 @@
+const User = require("../models/userModel");
+
+const getAllDoctorsInfo = async () => {
+  const doctors = await User.find({ role: "doctor" })
+    .select("firstName lastName email role  specialty")
+    .exec();
+  return doctors;
+};
+
+const getSingleDoctorInfo = async (doctorId)=>{
+    const doctor = await User.findById(doctorId)
+    .select("firstName lastName email role  specialty")
+    .exec();
+    return doctor
+}
+module.exports = {
+    getAllDoctorsInfo,
+    getSingleDoctorInfo
+}
